Added ability to hit character
Updated line tracing to not hit the capsule and created a new collision channel and set the player character's mesh to use it, so the line trace hits the mesh, which makes shooting feel more accurate.
Added hit montages and ability to hit a character with the projectile.
Added ability to have a weapon fire be automatic or not.
Adjusted the character rotation so it looks smooth on simulated proxies.