Added player health and elimination
Removed the default GameMode files and created our own based on GameMode so it can have its own Player Eliminated functionality, among other things.
Added health to the player and added a health bar on the overlay which decreases when the player receives damage.
Added ability to eliminate players and play a montage when that happens.