Added Respawning and Defeats
Added ability for eliminated player character to respawn.
Added effects upon character death - the character now "dissolves" by applying a new material to them and is beamed up with a new particle system. The character also drops their weapon.
The defeats are recorded and displayed in the character's HUD.
Added a new PlayerState to control the Score and the Defeats.