Added weapon ammo and reloading
Added ability to reload the weapon: added player-carried ammo and weapon-carried ammo, added ability to move ammo from the player to the weapon when pressing a key ('r' in this case), added HUD elements to show all ammo that update when values are changed.
Added a combat state so the player cannot fire while reloading.
Replicated the combat state and the carried ammo.
Added reloading montage and sounds to be played when picking up a weapon or reloading it.