Added match states and countdown timers
Synced time between the client and the server and added a countdown to the Game Mode and UI while the game is in progress. Added a countdown timer for the Warmup state and added a Cooldown state that displays the match's winners.
Added a Game Mode that stores the highest-scoring players.