Added more weapon and projectile types
Added a rocket launcher weapon and new rocket projectiles.
Added a pistol weapon which uses a line trace instead of firing a projectile actor.
Added an SMG weapon which is also a HitScan weapon which uses a line trace.
Added a shotgun weapon and added the concept of Scatter to HitScan weapons.
Added a sniper rifle and a sniper rifle aiming scope HUD element.
Added a grenade launcher.