Added hand grenades that can be thrown
Added reload animations for all weapons.
Added a weapon outline effect so the player can more easily identify weapons to pick up.
Added Grenades that can be thrown. Added an attachment mesh to the player's hand as well as an object that is spawned and thrown when the player presses 't'. Added ability for them to do damage. Added a section to the HUD to show how many grenades the player has left.