Added Ammo and Buff pickups
Added a "shield" to the character, variables in the Blaster class and a health-style bar in the HUD. If the character is damaged the shield now decremented before the character's health.
Added the concept of Pickups. Added Ammo-type pickups and Buff-type pickups. Buff types include increasing health, increasing the shield, increasing jump height, and increasing run speed. Increasing health & shield values can be done over time, and can also be done instantaneously if the BP is set to have a Time of 0.f.
Added pickup spawners which spawn pickups from a specified array at a specified, randomized interval (assuming the previously-spawned pickup was picked up).