Added a secondary weapon and high ping warning
Added a secondary weapon to the character. The default socket is on the character's back. When the character picks up a second weapon and they already have one in their hands, the new weapon is picked up and attaches to their back. It gets into a different weaponstate: EWS_Secondary.
The player can swap between their two picked up weapons.
Added a high-ping indicator to the screen to let the player know when their ping is high.
Added ability for player to see their shots as they fire locally rather than waiting for the server to multicast back to them.
Updated pickup widget to display it locally rather than waiting for the server to get back to the client before displaying it.