Adjusted Shotgun scatter calculations
Changed up the code so weapon scatter is not calculated on both the client and the server. That caused the scatter to be different in the two places, so the client could fire and look like they hit a character but since the server calculates a different scatter the server could believe that the player missed.
In the new way, only the client calculates scatter and then sends the FVector() hit data to the server, so both server and client believe the player hit roughly the same area in each shot, which will minimize aforementioned hit discrepancies.