Prevented laggy clients from getting stuck reloading
Added a bind to AnimInstance's OnMontageEnded delegate in order to prevent laggy clients from getting stuck in the reloading state.
For some reason the server would play the reload animation but never call the animnotify at the end that caused to character to enter the Unoccupied state. Using this delegate with an associated timer works more consistently than the animnotify.