Add weapon swapping animation
Disables using server-side rewind when the client's ping is too high. This is because we don't want to give players with good ping a bad experience.
Added an animation when swapping weapons. This includes similar backend timers for the couple of functions that need to be executed during the animation in order to prevent a laggy client from being stuck in the swapping weapon state. See the reload refactor for more details - the same issue happens with the swap anim.
Added validation for the fire delay timer which checks the value and kicks the player from the game if the value they sent is too different from the actual value.