Updated player elimination and top scoring player crown
Added animation for player being eliminated when a player chooses to leave the game.
Added telling the GameMode that a player left the game so that the TopScoringPlayers array can be adjusted accordingly.
Added a Crown to the Top Scoring player(s).
Refactored the PlayerGameState to prevent inappropriate access of it's array of Top Scoring Players.